Move `StdLoggerBackend` to a separate file
