+++ logging-backend/src/main/resources/META-INF/services/io.spine.logging.backend.system.BackendFactory
